Refactoring - deleted EndActivity.kt - end screen composable will be used instead - created endScreen Package and move endAnimation package into to it.
