+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,6 @@
         tools:targetApi="31">
 
         <activity android:name=".activities.GameActivity" />
-        <activity android:name=".activities.EndActivity" />
 
         <activity 
             android:name=".MainActivity"
